done creating booking with all 3 locking strategies

+++ src/main/java/com/example/AirbnbBookingSpring/models/readModels/AirbnbReadModel.java
@@ -22,7 +22,7 @@ public class AirbnbReadModel {
 
     private String location;
 
-    private Long pricePerNight;
+    private double pricePerNight;
 
     private List<AvailabilityReadModel> availability;
